cpp_m07 v1.1 cpp_m08 v0.0.1

+++ cpp_m07/ex02/Array.hpp
@@ -3,13 +3,15 @@
 
 #include <iostream>
 #include <stdexcept>
+#include <ctime>
+#include <cstdlib>
 
 template <typename T>
 class Array
 {
 private:
 	T *array;
-	size_t length;
+	unsigned int length;
 
 public:
 	Array();
